test(node-host): split runner tests under the max-lines limit
Separate lifecycle and connection tests from capability and inventory tests
so runner.test.ts stays below the max-lines gate blocking main CI. Share
the existing mocks and reset helper without changing tests or assertions.

Node-host before: 565 passed, 4 skipped (569 total), 48 files.
Node-host after: 565 passed, 4 skipped (569 total), 49 files.
Runner coverage remains 41 tests: 29 lifecycle/connection and 12 inventory.
