+++ src/node-host/runner.test.ts
@@ -1,255 +1,18 @@
-/** Tests node-host runner command parsing, timeout, and plugin dispatch behavior. */
-import { beforeEach, describe, expect, it, type Mock, vi } from "vitest";
+/** Tests node-host runner startup, connection configuration, and lifecycle. */
+import { beforeEach, describe, expect, it, vi } from "vitest";
 import { ConnectErrorDetailCodes } from "../../packages/gateway-protocol/src/connect-error-details.js";
-import { GATEWAY_SERVER_CAPS } from "../../packages/gateway-protocol/src/schema/frames.js";
 import { getConfigResolutionFacts, setConfigResolutionFacts } from "../config/resolution-facts.js";
 import type { GatewayClientOptions } from "../gateway/client.js";
 import {
-  NODE_RUNNER_INVENTORY_UPDATE_METHOD,
-  NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE,
-} from "../infra/node-runner-inventory.js";
-import type { configureNodeHost } from "./config.js";
-import { startNodeHostMcpManager, type NodeHostMcpManager } from "./mcp.js";
-import { runNodeHost } from "./runner.js";
-
-const NODE_PLUGIN_TOOLS_UPDATE_METHOD = "node.pluginTools.update";
-const NODE_SKILLS_UPDATE_METHOD = "node.skills.update";
-
-const mocks = vi.hoisted(() => ({
-  capturedGatewayClientOptions: [] as GatewayClientOptions[],
-  capturedConfiguredGatewayConfigs: [] as Array<{ contextPath?: string }>,
-  capturedGatewayClients: [] as Array<{
-    request: Mock<(method: string, params?: unknown) => Promise<unknown>>;
-    start: ReturnType<typeof vi.fn>;
-    stop: ReturnType<typeof vi.fn>;
-    updateNodeManifest: ReturnType<typeof vi.fn>;
-  }>,
-  mcpDescriptors: [] as Array<Record<string, unknown>>,
-  mcpDescriptorsChanged: undefined as (() => void) | undefined,
-  nodePluginTools: [] as Array<Record<string, unknown>>,
-  nodeSkillDescriptors: [] as Array<Record<string, unknown>>,
-  runtimeSteps: [] as string[],
-  useFakeRuntime: false,
-  fakeRuntimeWorkerHosting: false,
-  fakeRuntimeWorkerHostingDisabledReason: undefined as string | undefined,
-  runnerCapacityChanged: undefined as
-    | ((capacity: { total: number; available: number }) => void)
-    | undefined,
-  nodeHostCommands: [] as string[],
-  nodeHostCaps: [] as string[],
-  availabilityOnWatch: undefined as { caps: string[]; commands: string[] } | undefined,
-  availabilityChanged: undefined as (() => void) | undefined,
-  normalizedPath: null as string | null,
-  resolvedExecutables: new Map<string, string>(),
-  runtimeClient: undefined as
-    | { request: (method: string, params?: unknown) => Promise<unknown> }
-    | undefined,
-  closeMcpManager: vi.fn(async () => undefined),
-  runStartupMigrations: vi.fn(async () => undefined),
-  configureNodeHost: vi.fn(async (params: Parameters<typeof configureNodeHost>[0]) => {
-    mocks.capturedConfiguredGatewayConfigs.push(params.gateway);
-    return {
-      version: 1 as const,
-      nodeId: params.nodeId?.trim() || "node-test",
-      displayName: params.displayName?.trim() || params.fallbackDisplayName,
-      gateway: params.gateway,
-    };
-  }),
-  getRuntimeConfig: vi.fn<() => unknown>(() => ({ gateway: { handshakeTimeoutMs: 1_000 } })),
-  startGatewayClientWhenEventLoopReady: vi.fn(async () => ({
-    ready: false,
-    aborted: false,
-    elapsedMs: 0,
-  })),
-  resolveGatewayCredentialsWithSecretInputs: vi.fn(async (_params: { config: unknown }) => ({})),
-  activeRuntime: {
-    invoke: vi.fn(async () => {}),
-    handleInput: vi.fn(),
-    cancel: vi.fn(),
-    cancelAll: vi.fn(),
-    updateGatewayConnection: vi.fn(),
-    close: vi.fn(async () => {}),
-  },
-}));
-
-vi.mock("../config/config.js", () => ({
-  getRuntimeConfig: mocks.getRuntimeConfig,
-}));
-
-vi.mock("../gateway/client-start-readiness.js", () => ({
-  startGatewayClientWhenEventLoopReady: mocks.startGatewayClientWhenEventLoopReady,
-}));
-
-vi.mock("../gateway/client.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../gateway/client.js")>();
-  return {
-    ...actual,
-    GatewayClient: function GatewayClient(opts: GatewayClientOptions) {
-      const client = {
-        request: vi.fn(async () => ({})),
-        start: vi.fn(),
-        stop: vi.fn(),
-        updateNodeManifest: vi.fn(),
-      };
-      mocks.capturedGatewayClientOptions.push(opts);
-      mocks.capturedGatewayClients.push(client);
-      return client;
-    },
-  };
-});
-
-vi.mock("../gateway/credentials-secret-inputs.js", () => ({
-  resolveGatewayCredentialsWithSecretInputs: mocks.resolveGatewayCredentialsWithSecretInputs,
-}));
-
-vi.mock("../infra/device-identity.js", () => ({
-  loadOrCreateDeviceIdentity: vi.fn(() => ({
-    id: "device-test",
-    publicKey: "public-key-test",
-    privateKey: "private-key-test",
-  })),
-}));
-
-vi.mock("../infra/machine-name.js", () => ({
-  getMachineDisplayName: vi.fn(async () => "test-node"),
-}));
-
-vi.mock("../infra/machine-model.js", () => ({
-  resolveMachineModelIdentifier: vi.fn(() => "TestMachine1,1"),
-}));
-
-vi.mock("../infra/executable-path.js", () => ({
-  resolveExecutableFromPathEnv: vi.fn((bin: string) => mocks.resolvedExecutables.get(bin) ?? null),
-}));
-
-vi.mock("../infra/path-env.js", () => ({
-  ensureOpenClawCliOnPath: vi.fn(() => {
-    mocks.runtimeSteps.push("path");
-    if (mocks.normalizedPath) {
-      process.env.PATH = mocks.normalizedPath;
-    }
-  }),
-}));
-
-vi.mock("./config.js", () => ({
-  configureNodeHost: mocks.configureNodeHost,
-}));
-
-vi.mock("./plugin-node-host.js", () => ({
-  ensureNodeHostPluginRegistry: vi.fn(async () => undefined),
-  listRegisteredNodeHostCapsAndCommands: vi.fn((context: { env: NodeJS.ProcessEnv }) => {
-    mocks.runtimeSteps.push(`commands:${context.env.PATH ?? ""}`);
-    return {
-      commands: [...mocks.nodeHostCommands],
-      caps: [...mocks.nodeHostCaps],
-      nodePluginTools: [...mocks.nodePluginTools],
-    };
-  }),
-  watchRegisteredNodeHostCommandAvailability: vi.fn((_context: unknown, onChange: () => void) => {
-    mocks.availabilityChanged = onChange;
-    if (mocks.availabilityOnWatch) {
-      mocks.nodeHostCaps = [...mocks.availabilityOnWatch.caps];
-      mocks.nodeHostCommands = [...mocks.availabilityOnWatch.commands];
-    }
-    return () => {
-      mocks.availabilityChanged = undefined;
-    };
-  }),
-}));
-
-vi.mock("./mcp.js", () => ({
-  startNodeHostMcpManager: vi.fn(
-    async (
-      _servers: unknown,
-      deps?: {
-        onDescriptorsChanged?: () => void;
-      },
-    ) => {
-      mocks.mcpDescriptorsChanged = deps?.onDescriptorsChanged;
-      return {
-        descriptors: mocks.mcpDescriptors,
-        callMcpTool: vi.fn(),
-        close: mocks.closeMcpManager,
-      };
-    },
-  ),
-}));
-
-vi.mock("./skills.js", () => ({
-  scanNodeHostedSkills: vi.fn(() => mocks.nodeSkillDescriptors),
-}));
-
-vi.mock("./startup-state-migrations.js", () => ({
-  runStartupMigrations: mocks.runStartupMigrations,
-}));
-
-vi.mock("./runtime.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("./runtime.js")>();
-  return {
-    ...actual,
-    prepareNodeHostRuntime: async (
-      ...args: Parameters<typeof actual.prepareNodeHostRuntime>
-    ): ReturnType<typeof actual.prepareNodeHostRuntime> => {
-      if (!mocks.useFakeRuntime) {
-        return await actual.prepareNodeHostRuntime(...args);
-      }
-      return {
-        manifest: {
-          caps: [],
-          commands: [],
-          pathEnv: process.env.PATH ?? "",
-        },
-        workerHostingEnabled: mocks.fakeRuntimeWorkerHosting,
-        workerHostingDisabledReason: mocks.fakeRuntimeWorkerHostingDisabledReason,
-        initialInventory: { skills: [], pluginTools: [] },
-        start: (params) => {
-          mocks.runtimeClient = params.client;
-          mocks.runnerCapacityChanged = params.onRunnerCapacityChanged;
-          return mocks.activeRuntime;
-        },
-      };
-    },
-  };
-});
-
-function lastCapturedOptions(): GatewayClientOptions | undefined {
-  return mocks.capturedGatewayClientOptions.at(-1);
-}
+  lastCapturedOptions,
+  mocks,
+  resetRunnerTestState,
+  runNodeHost,
+  startNodeHostMcpManager,
+} from "./runner.test-support.js";
 
 describe("runNodeHost", () => {
-  beforeEach(() => {
-    mocks.capturedGatewayClientOptions.length = 0;
-    mocks.capturedConfiguredGatewayConfigs.length = 0;
-    mocks.capturedGatewayClients.length = 0;
-    mocks.mcpDescriptors = [];
-    mocks.mcpDescriptorsChanged = undefined;
-    mocks.nodePluginTools = [
-      {
-        pluginId: "test-plugin",
-        name: "remote_echo",
-        description: "Echo from node host",
-        command: "test.echo",
-        parameters: { type: "object", properties: {} },
-      },
-    ];
-    mocks.nodeSkillDescriptors = [];
-    mocks.runtimeSteps = [];
-    mocks.useFakeRuntime = false;
-    mocks.fakeRuntimeWorkerHosting = false;
-    mocks.fakeRuntimeWorkerHostingDisabledReason = undefined;
-    mocks.runnerCapacityChanged = undefined;
-    mocks.nodeHostCommands = [];
-    mocks.nodeHostCaps = [];
-    mocks.availabilityOnWatch = undefined;
-    mocks.availabilityChanged = undefined;
-    mocks.normalizedPath = null;
-    mocks.resolvedExecutables.clear();
-    mocks.runtimeClient = undefined;
-    vi.clearAllMocks();
-    mocks.getRuntimeConfig.mockReturnValue({
-      gateway: { handshakeTimeoutMs: 1_000 },
-    });
-  });
+  beforeEach(resetRunnerTestState);
 
   it("runs startup state migrations before constructing node-host state", async () => {
     await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
@@ -477,74 +240,6 @@ describe("runNodeHost", () => {
     });
   });
 
-  it("bootstraps PATH before probing plugin command availability", async () => {
-    const originalPath = process.env.PATH;
-    mocks.normalizedPath = "/normalized/node/path";
-    try {
-      await expect(
-        runNodeHost({
-          gatewayHost: "127.0.0.1",
-          gatewayPort: 18789,
-        }),
-      ).rejects.toThrow("event loop readiness timeout");
-    } finally {
-      process.env.PATH = originalPath;
-    }
-
-    expect(mocks.runtimeSteps).toEqual([
-      "path",
-      "commands:/normalized/node/path",
-      "commands:/normalized/node/path",
-    ]);
-  });
-
-  it("reconciles the manifest after watch attachment and on later changes", async () => {
-    mocks.startGatewayClientWhenEventLoopReady.mockResolvedValueOnce({
-      ready: true,
-      aborted: false,
-      elapsedMs: 0,
-    });
-    mocks.availabilityOnWatch = {
-      caps: ["canvas"],
-      commands: ["canvas.present"],
-    };
-    const processOnceSpy = vi.spyOn(process, "once");
-    const previousExitCode = process.exitCode;
-    try {
-      const running = runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 });
-      await vi.waitFor(() =>
-        expect(mocks.capturedGatewayClients[0]?.updateNodeManifest).toHaveBeenCalledWith(
-          expect.objectContaining({
-            caps: expect.arrayContaining(["canvas"]),
-            commands: expect.arrayContaining(["canvas.present"]),
-          }),
-        ),
-      );
-
-      mocks.nodeHostCaps = [];
-      mocks.nodeHostCommands = [];
-      mocks.availabilityChanged?.();
-      expect(mocks.capturedGatewayClients[0]?.updateNodeManifest).toHaveBeenLastCalledWith(
-        expect.objectContaining({
-          caps: expect.not.arrayContaining(["canvas"]),
-          commands: expect.not.arrayContaining(["canvas.present"]),
-        }),
-      );
-
-      const onSigterm = processOnceSpy.mock.calls.find(([event]) => event === "SIGTERM")?.[1];
-      onSigterm?.("SIGTERM");
-      await running;
-    } finally {
-      for (const [event, listener] of processOnceSpy.mock.calls) {
-        if ((event === "SIGINT" || event === "SIGTERM") && typeof listener === "function") {
-          process.off(event, listener);
-        }
-      }
-      process.exitCode = previousExitCode;
-      processOnceSpy.mockRestore();
-    }
-  });
-
   it("keeps a ref'd lifetime handle until a ready foreground host stops", async () => {
     mocks.startGatewayClientWhenEventLoopReady.mockResolvedValueOnce({
       ready: true,
@@ -617,351 +312,6 @@ describe("runNodeHost", () => {
       setIntervalSpy.mockRestore();
       clearIntervalSpy.mockRestore();
     }
-  });
-
-  it("declares the built-in MCP command family before any server is configured", async () => {
-    await expect(
-      runNodeHost({
-        gatewayHost: "127.0.0.1",
-        gatewayPort: 18789,
-      }),
-    ).rejects.toThrow("event loop readiness timeout");
-
-    expect(lastCapturedOptions()?.caps).toContain("mcp");
-    expect(lastCapturedOptions()?.commands).toContain("mcp.tools.call.v1");
-    expect(lastCapturedOptions()?.commands).not.toContain("agent.cli.claude.run.v1");
-    expect(lastCapturedOptions()?.workerRuns).toBeUndefined();
-  });
-
-  it("keeps unavailable worker hosting out of the handshake and reports the reason", async () => {
-    mocks.getRuntimeConfig.mockReturnValue({
-      gateway: { handshakeTimeoutMs: 1_000 },
-      nodeHost: { workerRuns: { enabled: true } },
-    } as never);
-    mocks.useFakeRuntime = true;
-    mocks.fakeRuntimeWorkerHostingDisabledReason = "Docker or Podman is unavailable";
-    const stderr = vi.spyOn(process.stderr, "write").mockImplementation(() => true);
-
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-
-    expect(lastCapturedOptions()?.workerRuns).toBeUndefined();
-    expect(stderr).toHaveBeenCalledExactlyOnceWith(
-      "node host worker hosting disabled: Docker or Podman is unavailable\n",
-    );
-    stderr.mockRestore();
-  });
-
-  it("advertises Claude agent runs only after node-local opt-in and binary resolution", async () => {
-    mocks.resolvedExecutables.set("claude", "/usr/bin/claude");
-    mocks.getRuntimeConfig.mockReturnValue({
-      gateway: { handshakeTimeoutMs: 1_000 },
-      nodeHost: { agentRuns: { claude: { enabled: true } } },
-    } as never);
-
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-
-    expect(lastCapturedOptions()?.commands).toContain("agent.cli.claude.run.v1");
-  });
-
-  it("publishes node plugin tools only after gateway hello succeeds", async () => {
-    await expect(
-      runNodeHost({
-        gatewayHost: "127.0.0.1",
-        gatewayPort: 18789,
-      }),
-    ).rejects.toThrow("event loop readiness timeout");
-
-    const options = mocks.capturedGatewayClientOptions[0];
-    const client = mocks.capturedGatewayClients[0];
-    expect(client?.request).not.toHaveBeenCalled();
-
-    options?.onHelloOk?.({
-      protocol: 1,
-      features: {
-        methods: [NODE_PLUGIN_TOOLS_UPDATE_METHOD],
-        events: [],
-      },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-
-    expect(client?.request).toHaveBeenCalledWith("node.pluginTools.update", {
-      tools: [
-        {
-          pluginId: "test-plugin",
-          name: "remote_echo",
-          description: "Echo from node host",
-          command: "test.echo",
-          parameters: { type: "object", properties: {} },
-        },
-      ],
-    });
-    expect(client?.request).toHaveBeenCalledWith(NODE_RUNNER_INVENTORY_UPDATE_METHOD, {
-      protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
-      workerHost: { enabled: false },
-    });
-  });
-
-  it("publishes opt-in consent and capacity in the atomic runner inventory", async () => {
-    mocks.getRuntimeConfig.mockReturnValue({
-      gateway: { handshakeTimeoutMs: 1_000 },
-      nodeHost: { workerRuns: { enabled: true, capacity: 5 } },
-    } as never);
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-    const options = mocks.capturedGatewayClientOptions[0];
-    const client = mocks.capturedGatewayClients[0];
-
-    options?.onHelloOk?.({
-      protocol: 4,
-      features: { methods: [], events: [] },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-
-    expect(client?.request).toHaveBeenCalledWith(NODE_RUNNER_INVENTORY_UPDATE_METHOD, {
-      protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
-      workerHost: {
-        enabled: true,
-        capacity: { total: 5, available: 5 },
-        bundlePrewarm: 1,
-      },
-    });
-  });
-
-  it("publishes each exact worker slot transition without reconnecting", async () => {
-    mocks.useFakeRuntime = true;
-    mocks.fakeRuntimeWorkerHosting = true;
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-    const options = mocks.capturedGatewayClientOptions[0];
-    const client = mocks.capturedGatewayClients[0];
-    expect(options?.workerRuns).toBeUndefined();
-
-    mocks.runnerCapacityChanged?.({ total: 2, available: 2 });
-    options?.onHelloOk?.({
-      protocol: 4,
-      features: {
-        methods: [],
-        events: [],
-        capabilities: [GATEWAY_SERVER_CAPS.NODE_WORKER_BUNDLE_RETENTION],
-      },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-    await vi.waitFor(() => {
-      expect(client?.request).toHaveBeenCalledWith(NODE_RUNNER_INVENTORY_UPDATE_METHOD, {
-        protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
-        workerHost: {
-          enabled: true,
-          capacity: { total: 2, available: 2 },
-          bundlePrewarm: 1,
-          bundleRetention: 1,
-        },
-      });
-    });
-
-    const negotiatedWorkerHost = {
-      enabled: true,
-      capacity: { total: 2, available: 2 },
-      bundlePrewarm: 1,
-      bundleRetention: 1,
-      bundleStatus: 1,
-      portalStream: 1,
-      environmentSession: 1,
-    };
-    options?.onHelloOk?.({
-      protocol: 4,
-      features: {
-        methods: [],
-        events: [],
-        capabilities: [
-          GATEWAY_SERVER_CAPS.NODE_WORKER_BUNDLE_RETENTION,
-          GATEWAY_SERVER_CAPS.NODE_WORKER_BUNDLE_STATUS,
-          GATEWAY_SERVER_CAPS.NODE_WORKER_PORTAL_STREAM,
-          GATEWAY_SERVER_CAPS.NODE_WORKER_ENVIRONMENT_SESSION,
-        ],
-      },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-    await vi.waitFor(() => {
-      expect(client?.request).toHaveBeenCalledWith(NODE_RUNNER_INVENTORY_UPDATE_METHOD, {
-        protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
-        workerHost: negotiatedWorkerHost,
-      });
-    });
-
-    const expectPublishedSlots = async (available: number) => {
-      mocks.runnerCapacityChanged?.({ total: 2, available });
-      await vi.waitFor(() => {
-        expect(client?.request).toHaveBeenLastCalledWith(NODE_RUNNER_INVENTORY_UPDATE_METHOD, {
-          protocolFeatures: [NODE_WORKER_SUPERVISOR_PROTOCOL_FEATURE],
-          workerHost: {
-            ...negotiatedWorkerHost,
-            capacity: { total: 2, available },
-          },
-        });
-      });
-    };
-    for (const available of [1, 0, 2]) {
-      await expectPublishedSlots(available);
-    }
-    expect(client?.updateNodeManifest).not.toHaveBeenCalled();
-  });
-
-  it("clears gateway plugin tools when the final node-hosted tool disappears", async () => {
-    mocks.startGatewayClientWhenEventLoopReady.mockResolvedValueOnce({
-      ready: true,
-      aborted: false,
-      elapsedMs: 0,
-    });
-    const processOnceSpy = vi.spyOn(process, "once");
-    const previousExitCode = process.exitCode;
-    try {
-      const running = runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 });
-      await vi.waitFor(() => expect(mocks.availabilityChanged).toBeDefined());
-      const client = mocks.capturedGatewayClients[0];
-      lastCapturedOptions()?.onHelloOk?.({
-        protocol: 1,
-        features: { methods: [NODE_PLUGIN_TOOLS_UPDATE_METHOD], events: [] },
-      } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-      expect(client?.request).toHaveBeenCalledWith("node.pluginTools.update", {
-        tools: [expect.objectContaining({ name: "remote_echo" })],
-      });
-
-      mocks.nodePluginTools = [];
-      mocks.availabilityChanged?.();
-
-      await vi.waitFor(() => {
-        expect(client?.request).toHaveBeenLastCalledWith("node.pluginTools.update", { tools: [] });
-      });
-      const onSigterm = processOnceSpy.mock.calls.find(([event]) => event === "SIGTERM")?.[1];
-      onSigterm?.("SIGTERM");
-      await running;
-    } finally {
-      for (const [event, listener] of processOnceSpy.mock.calls) {
-        if ((event === "SIGINT" || event === "SIGTERM") && typeof listener === "function") {
-          process.off(event, listener);
-        }
-      }
-      process.exitCode = previousExitCode;
-      processOnceSpy.mockRestore();
-    }
-  });
-
-  it("publishes node-hosted skills after gateway hello succeeds", async () => {
-    mocks.nodeSkillDescriptors = [
-      {
-        name: "release-helper",
-        description: "Prepare a release",
-        content: "---\nname: release-helper\ndescription: Prepare a release\n---\n",
-      },
-    ];
-
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-
-    const options = lastCapturedOptions();
-    expect(mocks.capturedGatewayClients[0]?.request).not.toHaveBeenCalledWith(
-      "node.skills.update",
-      expect.anything(),
-    );
-    options?.onHelloOk?.({
-      protocol: 1,
-      features: { methods: [NODE_SKILLS_UPDATE_METHOD], events: [] },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-    expect(mocks.capturedGatewayClients[0]?.request).toHaveBeenCalledWith("node.skills.update", {
-      skills: mocks.nodeSkillDescriptors,
-    });
-  });
-
-  it("does not publish node-hosted skills when disabled", async () => {
-    mocks.getRuntimeConfig.mockReturnValue({
-      gateway: { handshakeTimeoutMs: 1_000 },
-      nodeHost: { skills: { enabled: false } },
-    } as never);
-
-    await expect(runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 })).rejects.toThrow(
-      "event loop readiness timeout",
-    );
-    lastCapturedOptions()?.onHelloOk?.({
-      protocol: 1,
-      features: { methods: [NODE_SKILLS_UPDATE_METHOD], events: [] },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-
-    expect(mocks.capturedGatewayClients[0]?.request).not.toHaveBeenCalledWith(
-      "node.skills.update",
-      expect.anything(),
-    );
-  });
-
-  it("publishes plugin tools during MCP discovery and republishes catalog changes", async () => {
-    let resolveReadiness:
-      | ((value: { ready: false; aborted: false; elapsedMs: number }) => void)
-      | undefined;
-    mocks.startGatewayClientWhenEventLoopReady.mockReturnValueOnce(
-      new Promise((resolve) => {
-        resolveReadiness = resolve;
-      }),
-    );
-    let resolveManager: ((manager: NodeHostMcpManager) => void) | undefined;
-    vi.mocked(startNodeHostMcpManager).mockImplementationOnce(async (_servers, deps) => {
-      mocks.mcpDescriptorsChanged = deps?.onDescriptorsChanged;
-      return await new Promise((resolve) => {
-        resolveManager = resolve;
-      });
-    });
-    const running = runNodeHost({ gatewayHost: "127.0.0.1", gatewayPort: 18789 });
-    await vi.waitFor(() => expect(lastCapturedOptions()).toBeDefined());
-    lastCapturedOptions()?.onHelloOk?.({
-      protocol: 1,
-      features: { methods: [NODE_PLUGIN_TOOLS_UPDATE_METHOD], events: [] },
-    } as unknown as Parameters<NonNullable<GatewayClientOptions["onHelloOk"]>>[0]);
-    expect(mocks.capturedGatewayClients[0]?.request).toHaveBeenCalledWith(
-      "node.pluginTools.update",
-      { tools: [expect.objectContaining({ pluginId: "test-plugin" })] },
-    );
-
-    const descriptors: NodeHostMcpManager["descriptors"] = [
-      {
-        pluginId: "node-mcp",
-        name: "closed_search",
-        description: "Search closed server",
-        command: "mcp.tools.call.v1",
-        mcp: { server: "closed", tool: "search" },
-      },
-      {
-        pluginId: "node-mcp",
-        name: "healthy_search",
-        description: "Search healthy server",
-        command: "mcp.tools.call.v1",
-        mcp: { server: "healthy", tool: "search" },
-      },
-    ];
-    resolveManager?.({
-      descriptors,
-      callMcpTool: vi.fn(),
-      close: mocks.closeMcpManager,
-    });
-    const client = mocks.capturedGatewayClients[0];
-    const publishedToolNames = () => {
-      const params = client?.request.mock.calls.findLast(
-        ([method]) => method === NODE_PLUGIN_TOOLS_UPDATE_METHOD,
-      )?.[1] as { tools: Array<{ name?: string }> } | undefined;
-      return params?.tools.map((descriptor) => descriptor.name);
-    };
-    await vi.waitFor(() => {
-      expect(publishedToolNames()).toEqual(["closed_search", "healthy_search", "remote_echo"]);
-    });
-
-    descriptors.splice(0, 1);
-    expect(mocks.mcpDescriptorsChanged).toBeDefined();
-    mocks.mcpDescriptorsChanged?.();
-    await vi.waitFor(() => {
-      expect(publishedToolNames()).toEqual(["healthy_search", "remote_echo"]);
-    });
-    resolveReadiness?.({ ready: false, aborted: false, elapsedMs: 0 });
-    await expect(running).rejects.toThrow("event loop readiness timeout");
   });
 
   it.each([
